refactor: auto-split test_dbc_decorators.py
